Fix bug, the entire array is created every time a new book is created
Look at this incorrect code:
  bookShelf.textContent = "";
  myLibrary.map((book) => {
    createBook(book);
  });

Without bookShelf.textContent = "", the whole array myLibrary, is created. So if you have 5 elements in myLibrary, 5 books are created. However, with bookShelf.textContent = "", despite only one book being created, the books styling changes every time a new book is created, and its info (e.g. title, author, pages), becomes the same as the latest created book.

This block of code fixes the above issue as it gets the last iteration of myLibrary, and creates that:
  myLibrary.forEach(function (i, index, array) {
    if (index === array.length - 1) {
      createBook();
    }
  });